factor out plot thumbnail to container component
