+++ src/vs/workbench/contrib/positronPlots/browser/components/dynamicPlotThumbnail.tsx
@@ -4,15 +4,12 @@
 
 import * as React from 'react';
 import { useEffect, useState } from 'react'; // eslint-disable-line no-duplicate-imports
-import { PositronPlotsServices } from 'vs/workbench/contrib/positronPlots/browser/positronPlotsState';
 import { PlotClientInstance } from 'vs/workbench/services/languageRuntime/common/languageRuntimePlotClient';
 
 /**
  * DynamicPlotThumbnailProps interface.
  */
 interface DynamicPlotThumbnailProps {
-	selected: boolean;
-	plotService: PositronPlotsServices;
 	plotClient: PlotClientInstance;
 }
 
@@ -40,26 +37,14 @@ export const DynamicPlotThumbnail = (props: DynamicPlotThumbnailProps) => {
 		});
 	});
 
-	const selectPlot = () => {
-		props.plotService.positronPlotsService.selectPlot(props.plotClient.id);
-	};
-
-	const removePlot = () => {
-		props.plotService.positronPlotsService.removePlot(props.plotClient.id);
-	};
-
 	// If the plot is not yet rendered yet (no URI), show a placeholder;
 	// otherwise, show the rendered plot.
 	//
 	// Consider: we probably want a more explicit loading state; as written we
 	// will show the old URI until the new one is ready.
-	return (
-		<div className={'plot-thumbnail' + (props.selected ? ' selected' : '')}>
-			{uri && <button className='image-wrapper' onClick={selectPlot}>
-				<img src={uri} alt={'Plot ' + props.plotClient.id} />
-			</button>}
-			{!uri && <button className='plot-thumbnail-placeholder' onClick={selectPlot}></button>}
-			<button className='plot-close codicon codicon-close' onClick={removePlot}></button>
-		</div >
-	);
+	if (uri) {
+		return <img src={uri} alt={'Plot ' + props.plotClient.id} />;
+	} else {
+		return <div className='plot-thumbnail-placeholder'></div>;
+	}
 };
