+++ src/vs/workbench/contrib/positronPlots/browser/components/staticPlotThumbnail.tsx
@@ -3,15 +3,12 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as React from 'react';
-import { PositronPlotsServices } from 'vs/workbench/contrib/positronPlots/browser/positronPlotsState';
 import { StaticPlotClient } from 'vs/workbench/services/positronPlots/common/staticPlotClient';
 
 /**
  * StaticPlotThumbnailProps interface.
  */
 interface StaticPlotThumbnailProps {
-	selected: boolean;
-	plotService: PositronPlotsServices;
 	plotClient: StaticPlotClient;
 }
 
@@ -22,22 +19,5 @@ interface StaticPlotThumbnailProps {
  * @returns The rendered component.
  */
 export const StaticPlotThumbnail = (props: StaticPlotThumbnailProps) => {
-
-	const selectPlot = () => {
-		props.plotService.positronPlotsService.selectPlot(props.plotClient.id);
-	};
-
-	const removePlot = () => {
-		props.plotService.positronPlotsService.removePlot(props.plotClient.id);
-	};
-
-	return (
-		<div className={'plot-thumbnail' + (props.selected ? ' selected' : '')}>
-			<button className='image-wrapper'>
-				<img src={props.plotClient.uri} alt={'Plot ' + props.plotClient.id}
-					onClick={selectPlot} />
-			</button>
-			<button className='plot-close codicon codicon-close' onClick={removePlot}></button>
-		</div>
-	);
+	return <img src={props.plotClient.uri} alt={'Plot ' + props.plotClient.id} />;
 };
